better subscription polling
